css/css-flexbox/parsing/align-items-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing align-items with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-align-items">
<meta name="assert" content="align-items supports the full grammar 'flex-start | flex-end | center | baseline | stretch'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("align-items", "flex-start");
test_valid_value("align-items", "flex-end");
test_valid_value("align-items", "center");
test_valid_value("align-items", "baseline");
test_valid_value("align-items", "stretch");
</script>
</body>
</html>
